Move logback 12 version adapter to single module.
