+++ logger-adapter-impl/logback-adapter-12/src/main/java/com/alibaba/nacos/logger/adapter/logback12/NacosClientPropertyAction.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2022 Alibaba Group Holding Ltd.
+ * Copyright 1999-2023 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,20 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.client.logging.logback;
+package com.alibaba.nacos.logger.adapter.logback12;
 
 import ch.qos.logback.core.joran.action.Action;
 import ch.qos.logback.core.joran.action.ActionUtil;
 import ch.qos.logback.core.joran.spi.ActionException;
 import ch.qos.logback.core.joran.spi.InterpretationContext;
 import ch.qos.logback.core.util.OptionHelper;
-import com.alibaba.nacos.client.env.NacosClientProperties;
 import com.alibaba.nacos.common.logging.NacosLoggingProperties;
 import org.xml.sax.Attributes;
 
 /**
- * support logback read properties from NacosClientProperties. just like springProperty.
- * for example:
+ * support logback read properties from NacosClientProperties. just like springProperty. for example:
  * <nacosClientProperty scope="context" name="logPath" source="system.log.path" defaultValue="/root" />
+ *
  * @author onewe
  */
 class NacosClientPropertyAction extends Action {
