+++ logger-adapter-impl/logback-adapter-12/src/main/resources/META-INF/services/com.alibaba.nacos.common.logging.NacosLoggingAdapter
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-com.alibaba.nacos.client.logging.log4j2.Log4J2NacosLoggingAdapter
+com.alibaba.nacos.logger.adapter.logback12.LogbackNacosLoggingAdapter
